requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
fsspec==2021.5.0
future==0.18.2
google-auth==1.32.0
google-auth-oauthlib==0.4.4
grpcio==1.38.1
gunicorn==20.1.0
idna==2.10
Markdown==3.3.4
multidict==5.1.0
numpy==1.21.0
oauthlib==3.1.1
olefile==0.46
packaging==20.9
Pillow==8.2.0
protobuf==3.17.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.0
pyparsing==2.4.7
pytorch-lightning==1.3.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
torch==1.9.0
torchmetrics==0.3.2
torchvision==0.10.0
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
whitenoise==5.2.0
yarl==1.6.3
